+++ app/server/api.py
@@ -1,10 +1,10 @@
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
-from app.server.auth.authenticate import router as auth_router
-from app.server.routes.data import router as data_router
-from app.server.routes.policies import router as api_router
-from app.server.routes.user import router as user_router
+from app.server.auth.get_token import router as auth_router
+from app.server.routes.data_routes import router as data_router
+from app.server.routes.policy_routes import router as api_router
+from app.server.routes.repo_routes import router as user_router
 
 app = FastAPI(
     swagger_ui_parameters={"syntaxHighlight.theme": "obsidian"},
