+++ app/utils/write_rego.py
@@ -1,6 +1,6 @@
-from app.server.services.github import GitHubOperations
+from app.server.services.github_operations import GitHubOperations
 
-from ..server.services.gitlab import GitLabOperations
+from ..server.services.gitlab_operations import GitLabOperations
 from .build_rego_file import build_rego
 
 initiate_rule = "package httpapi.authz\nimport input\ndefault allow = false\n\n\n\n"
